refactor: abstract apollo context
