+++ context/ApolloContext.jsx
@@ -1,0 +1,30 @@
+import {
+  ApolloClient,
+  ApolloProvider,
+  createHttpLink,
+  InMemoryCache,
+} from "@apollo/client";
+import { setContext } from "@apollo/client/link/context";
+
+const httpLink = createHttpLink({
+  uri: "https://api-gateway-bujosa.cloud.okteto.net/graphql",
+});
+
+const authLink = setContext((_, { headers }) => {
+  const accessToken = localStorage.getItem("accessToken");
+  return {
+    headers: {
+      ...headers,
+      authorization: accessToken ? `Bearer ${accessToken}` : "",
+    },
+  };
+});
+
+const graphQLClient = new ApolloClient({
+  link: authLink.concat(httpLink),
+  cache: new InMemoryCache(),
+});
+
+export function GraphQLProvider({ children }) {
+  return <ApolloProvider client={graphQLClient}>{children}</ApolloProvider>;
+}
