+++ pages/_app.js
@@ -1,37 +1,12 @@
-import {
-  ApolloClient,
-  ApolloProvider,
-  createHttpLink,
-  InMemoryCache,
-} from "@apollo/client";
-import { setContext } from "@apollo/client/link/context";
 import { ChakraProvider } from "@chakra-ui/react";
 import Layout from "../components/layout/Layout";
+import { GraphQLProvider } from "../context/ApolloContext";
 import { AuthProvider } from "../context/AuthContext";
 import "../styles/globals.css";
 
-const httpLink = createHttpLink({
-  uri: "https://api-gateway-bujosa.cloud.okteto.net/graphql",
-});
-
-const authLink = setContext((_, { headers }) => {
-  const token = localStorage.getItem("token");
-  return {
-    headers: {
-      ...headers,
-      authorization: token ? `Bearer ${token}` : "",
-    },
-  };
-});
-
-const graphQLClient = new ApolloClient({
-  link: authLink.concat(httpLink),
-  cache: new InMemoryCache(),
-});
-
 function MyApp({ Component, pageProps }) {
   return (
-    <ApolloProvider client={graphQLClient}>
+    <GraphQLProvider>
       <AuthProvider>
         <ChakraProvider>
           <Layout>
@@ -39,7 +14,7 @@ function MyApp({ Component, pageProps }) {
           </Layout>
         </ChakraProvider>
       </AuthProvider>
-    </ApolloProvider>
+    </GraphQLProvider>
   );
 }
